Seeders terminados: seeders y factories de Genérico, Pedido, Genérico Pedido, Genérico Pedido Producto, Internacion, Dirección Proveedor. Ajustes en las migraciones

+++ database/migrations/2021_10_13_024446_create_supplier_addresses_table.php
@@ -17,7 +17,7 @@ class CreateSupplierAddressesTable extends Migration
             $table->id();
             $table->string('zip_code');
             $table->string('street');
-            $table->smallInteger('number');
+            $table->integer('number');
             $table->timestamps();
         });
     }
